refactor: make Swiper client-only and upgrade
