+++ var/www/frontend-next/package.json
@@ -15,7 +15,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.10.1",
-    "swiper": "^9.4.1",
+    "swiper": "^11",
     "zustand": "4.4.0"
   },
   "devDependencies": {
